bench: add flipped join merge benchmark
